# openapi.yaml
openapi: 3.0.0
info:
  title: Another Home API
  description: The comprehensive microservices API documentation for the Another Home Hostel Management System.
  version: 1.1.0
...
paths:
...
  /accommodation/rooms:
    get:
      tags:
        - Accommodation
      summary: Get all rooms and their current capacity
      responses:
        "200":
          description: Returns an array of all rooms.
    post:
      tags:
        - Accommodation
      summary: Create a new hostel room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomDto"
      responses:
        "201":
          description: Room created successfully.
  /accommodation/allocations:
    post:
      tags:
        - Accommodation
      summary: Assign a student to a specific bed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocateBedDto"
      responses:
        "201":
          description: Student successfully allocated.

  # ==========================================
  # 3. OPERATIONS (Operations Microservice)
  # ==========================================
  # --- Maintenance ---
...
components:
  schemas:
...
    CreateRoomDto:
      type: object
      required:
        - roomNumber
        - capacity
        - designation
      properties:
        roomNumber:
          type: string
          example: A-101
        capacity:
          type: integer
          example: 4
        designation:
          type: string
          enum: [Male, Female, Neutral]
          example: Male
    AllocateBedDto:
      type: object
      required:
        - studentId
        - bedId
      properties:
        studentId:
          type: string
          example: STU-2023-045
        bedId:
          type: string
          example: BED-A101-01

    # --- Operations Schemas ---
...